+++ admission/rules/v1/r2000_exec_to_pod.go
@@ -4,9 +4,11 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/kubescape/go-logger"
+	"github.com/kubescape/go-logger/helpers"
 	"github.com/kubescape/operator/admission/rules"
 	"github.com/kubescape/operator/objectcache"
-	"go.uber.org/zap"
+
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apiserver/pkg/admission"
 	"k8s.io/apiserver/pkg/authentication/user"
@@ -69,12 +71,11 @@ func (rule *R2000ExecToPod) ProcessEvent(event admission.Attributes, access inte
 
 	client := access.(objectcache.KubernetesCache).GetClientset()
 
-	workloadKind, workloadName, workloadNamespace, nodeName, err := GetParentWorkloadDetails(event, client)
+	workloadKind, workloadName, workloadNamespace, nodeName, err := GetControllerDetails(event, client)
 	if err != nil {
-		zap.L().Error("Failed to get parent workload details", zap.Error(err))
+		logger.L().Error("Failed to get parent workload details", helpers.Error(err))
 		return nil
 	}
-
 
 	ruleFailure := GenericRuleFailure{
 		BaseRuntimeAlert: apitypes.BaseRuntimeAlert{
@@ -106,12 +107,12 @@ func (rule *R2000ExecToPod) ProcessEvent(event admission.Attributes, access inte
 			RuleDescription: fmt.Sprintf("Exec to pod detected on pod %s", event.GetName()),
 		},
 		RuntimeAlertK8sDetails: apitypes.RuntimeAlertK8sDetails{
-			PodName:      event.GetName(),
-			Namespace:    event.GetNamespace(),
-			WorkloadName: workloadName,
+			PodName:           event.GetName(),
+			Namespace:         event.GetNamespace(),
+			WorkloadName:      workloadName,
 			WorkloadNamespace: workloadNamespace,
-			WorkloadKind: workloadKind,
-			NodeName:     nodeName,
+			WorkloadKind:      workloadKind,
+			NodeName:          nodeName,
 		},
 		RuleID: R2000ID,
 	}
