+++ admission/rules/v1/r2001_portforward.go
@@ -6,7 +6,9 @@ import (
 
 	"github.com/kubescape/operator/admission/rules"
 	"github.com/kubescape/operator/objectcache"
-	"go.uber.org/zap"
+
+	"github.com/kubescape/go-logger"
+	"github.com/kubescape/go-logger/helpers"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apiserver/pkg/admission"
 	"k8s.io/apiserver/pkg/authentication/user"
@@ -69,9 +71,9 @@ func (rule *R2001PortForward) ProcessEvent(event admission.Attributes, access in
 
 	client := access.(objectcache.KubernetesCache).GetClientset()
 
-	workloadKind, workloadName, workloadNamespace, nodeName, err := GetParentWorkloadDetails(event, client)
+	workloadKind, workloadName, workloadNamespace, nodeName, err := GetControllerDetails(event, client)
 	if err != nil {
-		zap.L().Error("Failed to get parent workload details", zap.Error(err))
+		logger.L().Error("Failed to get parent workload details", helpers.Error(err))
 		return nil
 	}
 
@@ -105,12 +107,12 @@ func (rule *R2001PortForward) ProcessEvent(event admission.Attributes, access in
 			RuleDescription: fmt.Sprintf("Port forward detected on pod %s", event.GetName()),
 		},
 		RuntimeAlertK8sDetails: apitypes.RuntimeAlertK8sDetails{
-			PodName:      event.GetName(),
-			Namespace:    event.GetNamespace(),
-			WorkloadName: workloadName,
+			PodName:           event.GetName(),
+			Namespace:         event.GetNamespace(),
+			WorkloadName:      workloadName,
 			WorkloadNamespace: workloadNamespace,
-			WorkloadKind: workloadKind,
-			NodeName:     nodeName,
+			WorkloadKind:      workloadKind,
+			NodeName:          nodeName,
 		},
 		RuleID: R2001ID,
 	}
